django/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
flake8==3.8.4
psycopg2==2.8.6
Pillow==8.0.1
django-ckeditor==6.0.0
flake8==3.8.4
black==20.8b1